% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/decrements.r
\docType{package}
\name{decrements}
\alias{decrements}
\alias{decrements-package}
\title{Decrement tables for analysis of public pension plans.}
\description{
Currently includes:\cr
\code{\link{mortality}} Various mortality tables in a single data frame\cr
\code{\link{termination}} Various termination tables in a single data frame\cr
}

