% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mortality.r
\docType{data}
\name{mortality}
\alias{mortality}
\title{Mortality decrement tables.}
\format{Data frame with 1 row per age per tablename.
\describe{
\item{tablename}{Unique name identifying the mortality table - ordinarily will select on this, character}
\item{tid}{SOA unique table number if this table is on the SOA site (rather than constructed), integer}
\item{series}{Level of government: 1, 2, or 3, for State-local, State, or local, numeric}
\item{memtype}{Purpose of the table: c("employee", "annuitant", "disabled", "hybrid")}
\item{collar}{c("allcollars", "blue", "white")}
\item{sex}{Male or female, c("male", "female", "unisex", "female75"), character}
      unisex is 50% male, 50% female
      female75 is 75% female, 25% male
\item{age}{Attained age, integer}
\item{qxm}{Rate of mortality at age x, numeric}
}}
\source{
Society of Actuaries (http://mort.soa.org/)\cr
}
\usage{
mortality
}
\description{
Selected decrement tables.
}
\examples{
library(dplyr)
glimpse(mortality)
count(mortality, tablename)
}
\keyword{datasets}

